==> 3_s_m19.ipynb <==
#### MALDI-TOF ANALISIS VINCHUCAS
##### 3) s_m19

##### s: Supervisado
##### m19:  Utiliza 19 muestras correspondientes a cada individuo

##### Autor: Bioing. Facundo Urteaga (IBB-CONICET)

# Carga de Librerías

import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
from sklearn import datasets, neighbors
from mlxtend.plotting import plot_decision_regions
import seaborn as sns
from sklearn.linear_model import LogisticRegression  # Modelo de regresión logística

# Carga de datos

df_19_top5 = pd.read_csv('matint_19_top5.csv')
metadata_19_top5 = pd.read_csv('df_19.csv')
